refactor(schema): 修改 Entity 的内部 ID 类型为 std::size_t

+++ include/dreamdb/schema/collection.h
@@ -98,7 +98,7 @@ public:
      * @param id 要删除的实体的 ID
      * @return 删除结果
      */
-    MutationResult remove(std::int64_t id);
+    MutationResult remove(std::size_t id);
 
     /**
      * @brief 更新实体
@@ -106,7 +106,7 @@ public:
      * @param fields 要更新的字段列表
      * @return 更新结果
      */
-    MutationResult update(std::int64_t id, std::vector<std::pair<std::size_t, FieldValue>> fields);
+    MutationResult update(std::size_t id, std::vector<std::pair<std::size_t, FieldValue>> fields);
 
     /**
      * @brief 查询实体
@@ -119,7 +119,7 @@ private:
     std::string name_;                                  // 集合名称
     std::vector<Field> schema_;                         // 字段定义列表
     std::unordered_map<std::string, std::size_t> field_index_map_; // 字段索引映射
-    std::int64_t next_id_;                              // 自增 ID 生成器
+    std::size_t next_id_;                               // 自增 ID 生成器
     std::unique_ptr<SegmentManager> segment_manager_;   // 段管理器
     std::shared_ptr<Segment> active_segment_;           // 当前活动段
 };
